+++ itest/types/datagen.go
@@ -10,8 +10,8 @@ import (
 	"github.com/babylonchain/babylon/crypto/eots"
 	"github.com/babylonchain/babylon/testutil/datagen"
 	bbn "github.com/babylonchain/babylon/types"
-	"github.com/babylonchain/babylon/x/btcstaking/types"
 	ftypes "github.com/babylonchain/babylon/x/finality/types"
+	fptypes "github.com/babylonchain/finality-provider/types"
 	"github.com/btcsuite/btcd/btcec/v2"
 	"github.com/btcsuite/btcd/btcec/v2/schnorr"
 	"github.com/btcsuite/btcd/chaincfg"
@@ -21,110 +21,9 @@ import (
 	bstypes "github.com/babylonchain/babylon/x/btcstaking/types"
 )
 
-type NewFinalityProvider struct {
-	Description *FinalityProviderDescription `json:"description,omitempty"`
-	Commission  string                       `json:"commission"`
-	BabylonPK   *PubKey                      `json:"babylon_pk,omitempty"`
-	BTCPKHex    string                       `json:"btc_pk_hex"`
-	Pop         *ProofOfPossession           `json:"pop,omitempty"`
-	ConsumerID  string                       `json:"consumer_id"`
-}
-
-type FinalityProviderDescription struct {
-	Moniker         string `json:"moniker"`
-	Identity        string `json:"identity"`
-	Website         string `json:"website"`
-	SecurityContact string `json:"security_contact"`
-	Details         string `json:"details"`
-}
-
-type PubKey struct {
-	Key string `json:"key"`
-}
-
-type ProofOfPossession struct {
-	BTCSigType int32  `json:"btc_sig_type"`
-	BabylonSig string `json:"babylon_sig"`
-	BTCSig     string `json:"btc_sig"`
-}
-
-type CovenantAdaptorSignatures struct {
-	CovPK       string   `json:"cov_pk"`
-	AdaptorSigs []string `json:"adaptor_sigs"`
-}
-
-type SignatureInfo struct {
-	PK  string `json:"pk"`
-	Sig string `json:"sig"`
-}
-
-type BtcUndelegationInfo struct {
-	UnbondingTx           string                      `json:"unbonding_tx"`
-	DelegatorUnbondingSig string                      `json:"delegator_unbonding_sig"`
-	CovenantUnbondingSigs []SignatureInfo             `json:"covenant_unbonding_sig_list"`
-	SlashingTx            string                      `json:"slashing_tx"`
-	DelegatorSlashingSig  string                      `json:"delegator_slashing_sig"`
-	CovenantSlashingSigs  []CovenantAdaptorSignatures `json:"covenant_slashing_sigs"`
-}
-
-type ActiveBtcDelegation struct {
-	BTCPkHex             string                      `json:"btc_pk_hex"`
-	FpBtcPkList          []string                    `json:"fp_btc_pk_list"`
-	StartHeight          uint64                      `json:"start_height"`
-	EndHeight            uint64                      `json:"end_height"`
-	TotalSat             uint64                      `json:"total_sat"`
-	StakingTx            string                      `json:"staking_tx"`
-	SlashingTx           string                      `json:"slashing_tx"`
-	DelegatorSlashingSig string                      `json:"delegator_slashing_sig"`
-	CovenantSigs         []CovenantAdaptorSignatures `json:"covenant_sigs"`
-	StakingOutputIdx     uint32                      `json:"staking_output_idx"`
-	UnbondingTime        uint32                      `json:"unbonding_time"`
-	UndelegationInfo     BtcUndelegationInfo         `json:"undelegation_info"`
-	ParamsVersion        uint32                      `json:"params_version"`
-}
-
-type SlashedBtcDelegation struct {
-	// Define fields as needed
-}
-
-type UnbondedBtcDelegation struct {
-	// Define fields as needed
-}
-
-type BtcStaking struct {
-	NewFP       []NewFinalityProvider   `json:"new_fp"`
-	ActiveDel   []ActiveBtcDelegation   `json:"active_del"`
-	SlashedDel  []SlashedBtcDelegation  `json:"slashed_del"`
-	UnbondedDel []UnbondedBtcDelegation `json:"unbonded_del"`
-}
-
-type CommitPublicRandomness struct {
-	FPPubKeyHex string `json:"fp_pubkey_hex"`
-	StartHeight uint64 `json:"start_height"`
-	NumPubRand  uint64 `json:"num_pub_rand"`
-	Commitment  string `json:"commitment"`
-	Signature   string `json:"signature"`
-}
-
-type SubmitFinalitySignature struct {
-	FpPubkeyHex string `json:"fp_pubkey_hex"`
-	Height      uint64 `json:"height"`
-	PubRand     string `json:"pub_rand"`   // base64 encoded
-	Proof       Proof  `json:"proof"`      // nested struct
-	BlockHash   string `json:"block_hash"` // base64 encoded
-	Signature   string `json:"signature"`  // base64 encoded
-}
-
-type Proof struct {
-	Total    uint64   `json:"total"`
-	Index    uint64   `json:"index"`
-	LeafHash string   `json:"leaf_hash"`
-	Aunts    []string `json:"aunts"`
-}
-
-func genRandomFinalityProvider() NewFinalityProvider {
-	return NewFinalityProvider{
-		Description: &FinalityProviderDescription{
+func genRandomFinalityProvider() fptypes.NewFinalityProvider {
+	return fptypes.NewFinalityProvider{
+		Description: &fptypes.FinalityProviderDescription{
 			Moniker:         "fp1",
 			Identity:        "Finality Provider 1",
 			Website:         "https://fp1.com",
@@ -132,11 +31,11 @@ func genRandomFinalityProvider() NewFinalityProvider {
 			Details:         "details",
 		},
 		Commission: "0.05",
-		BabylonPK: &PubKey{
+		BabylonPK: &fptypes.PubKey{
 			Key: base64.StdEncoding.EncodeToString([]byte("mock_pub_rand")),
 		},
 		BTCPKHex: "1",
-		Pop: &ProofOfPossession{
+		Pop: &fptypes.ProofOfPossession{
 			BTCSigType: 0,
 			BabylonSig: base64.StdEncoding.EncodeToString([]byte("mock_babylon_sig")),
 			BTCSig:     base64.StdEncoding.EncodeToString([]byte("mock_btc_sig")),
@@ -145,7 +44,7 @@ func genRandomFinalityProvider() NewFinalityProvider {
 	}
 }
 
-func genRandomBtcDelegation() (*types.Params, ActiveBtcDelegation) {
+func genRandomBtcDelegation() (*bstypes.Params, fptypes.ActiveBtcDelegation) {
 	var net = &chaincfg.RegressionNetParams
 	r := rand.New(rand.NewSource(time.Now().Unix()))
 	t := &testing.T{}
@@ -173,7 +72,7 @@ func genRandomBtcDelegation() (*types.Params, ActiveBtcDelegation) {
 	unbondingTime := uint16(100) + 1
 	slashingChangeLockTime := unbondingTime
 
-	bsParams := &types.Params{
+	bsParams := &bstypes.Params{
 		CovenantPks:     bbn.NewBIP340PKsFromBTCPKs(covenantPKs),
 		CovenantQuorum:  covenantQuorum,
 		SlashingAddress: slashingAddress.EncodeAddress(),
@@ -205,45 +104,45 @@ func genRandomBtcDelegation() (*types.Params, ActiveBtcDelegation) {
 	return bsParams, activeDel
 }
 
-func convertBTCDelegationToActiveBtcDelegation(mockDel *bstypes.BTCDelegation) ActiveBtcDelegation {
+func convertBTCDelegationToActiveBtcDelegation(mockDel *bstypes.BTCDelegation) fptypes.ActiveBtcDelegation {
 	var fpBtcPkList []string
 	for _, pk := range mockDel.FpBtcPkList {
 		fpBtcPkList = append(fpBtcPkList, pk.MarshalHex())
 	}
 
-	var covenantSigs []CovenantAdaptorSignatures
+	var covenantSigs []fptypes.CovenantAdaptorSignatures
 	for _, cs := range mockDel.CovenantSigs {
 		var adaptorSigs []string
 		for _, sig := range cs.AdaptorSigs {
 			adaptorSigs = append(adaptorSigs, base64.StdEncoding.EncodeToString(sig))
 		}
-		covenantSigs = append(covenantSigs, CovenantAdaptorSignatures{
+		covenantSigs = append(covenantSigs, fptypes.CovenantAdaptorSignatures{
 			CovPK:       cs.CovPk.MarshalHex(),
 			AdaptorSigs: adaptorSigs,
 		})
 	}
 
-	var covenantUnbondingSigs []SignatureInfo
+	var covenantUnbondingSigs []fptypes.SignatureInfo
 	for _, sigInfo := range mockDel.BtcUndelegation.CovenantUnbondingSigList {
-		covenantUnbondingSigs = append(covenantUnbondingSigs, SignatureInfo{
+		covenantUnbondingSigs = append(covenantUnbondingSigs, fptypes.SignatureInfo{
 			PK:  sigInfo.Pk.MarshalHex(),
 			Sig: base64.StdEncoding.EncodeToString(sigInfo.Sig.MustMarshal()),
 		})
 	}
 
-	var covenantSlashingSigs []CovenantAdaptorSignatures
+	var covenantSlashingSigs []fptypes.CovenantAdaptorSignatures
 	for _, cs := range mockDel.BtcUndelegation.CovenantSlashingSigs {
 		var adaptorSigs []string
 		for _, sig := range cs.AdaptorSigs {
 			adaptorSigs = append(adaptorSigs, base64.StdEncoding.EncodeToString(sig))
 		}
-		covenantSlashingSigs = append(covenantSlashingSigs, CovenantAdaptorSignatures{
+		covenantSlashingSigs = append(covenantSlashingSigs, fptypes.CovenantAdaptorSignatures{
 			CovPK:       cs.CovPk.MarshalHex(),
 			AdaptorSigs: adaptorSigs,
 		})
 	}
 
-	undelegationInfo := BtcUndelegationInfo{
+	undelegationInfo := fptypes.BtcUndelegationInfo{
 		UnbondingTx:           base64.StdEncoding.EncodeToString(mockDel.BtcUndelegation.UnbondingTx),
 		SlashingTx:            base64.StdEncoding.EncodeToString(mockDel.BtcUndelegation.SlashingTx.MustMarshal()),
 		DelegatorSlashingSig:  base64.StdEncoding.EncodeToString(mockDel.BtcUndelegation.DelegatorSlashingSig.MustMarshal()),
@@ -251,7 +150,7 @@ func convertBTCDelegationToActiveBtcDelegation(mockDel *bstypes.BTCDelegation) A
 		CovenantSlashingSigs:  covenantSlashingSigs,
 	}
 
-	return ActiveBtcDelegation{
+	return fptypes.ActiveBtcDelegation{
 		BTCPkHex:             mockDel.BtcPk.MarshalHex(),
 		FpBtcPkList:          fpBtcPkList,
 		StartHeight:          mockDel.StartHeight,
